api working for 1 table
This is a baby working example for 1 table on my db. I just need to add code for the other two tables. But, its time to put make up on and iron my hair. Saturday night and girls know it. It took more than I expected the os configuration and setup connection with sqlalchemy and postgres... :p